refactor(step2): LadderParser - 메서드 위치 변경(StringUtil 추가)
