+++ src/main/java/nextstep/ladder/util/StringUtil.java
@@ -1,2 +1,12 @@
-package nextstep.ladder.util;public class StringUtil {
+package nextstep.ladder.util;
+
+public class StringUtil {
+
+    public static String appendWhiteSpaceInside(String text, int length) {
+        return String.format(new StringBuilder()
+                .append("%")
+                .append(length)
+                .append("s")
+                .toString(), text);
+    }
 }
